+++ server/src/services/geolocation.service.ts
@@ -1,6 +1,6 @@
 import { internal, isBoom } from "@hapi/boom"
 import { AxiosResponse } from "axios"
-import { IPointFeature } from "shared/models/cacheGeolocation.model"
+import { IPointFeature, IPointGeometry, ZPointGeometry } from "shared/models/cacheGeolocation.model"
 
 import { getHttpClient } from "@/common/utils/httpUtils"
 
@@ -97,3 +97,47 @@ export const getGeolocation = async (rawAddress: string): Promise<IPointFeature 
     return null
   }
 }
+
+export const getGeoPoint = async (adresse: string): Promise<IPointGeometry> => {
+  try {
+    const geolocation = await getGeolocation(adresse)
+    if (!geolocation) {
+      throw internal("getGeoPoint: addresse non trouvée", { adresse })
+    }
+
+    return ZPointGeometry.parse(geolocation.geometry)
+  } catch (error: any) {
+    if (isBoom(error)) {
+      throw error
+    }
+    const newError = internal(`getGeoPoint: erreur de récupération des geo coordonnées`, { adresse })
+    newError.cause = error
+    throw newError
+  }
+}
+
+export type GeoCoord = {
+  latitude: number
+  longitude: number
+}
+
+export const getGeoCoordinates = async (adresse: string): Promise<GeoCoord> => {
+  const geopoint = await getGeoPoint(adresse)
+  const [longitude, latitude] = geopoint.coordinates
+
+  return { latitude, longitude }
+}
+
+export function geometryToGeoCoord(geometry): [number, number] {
+  const { type } = geometry
+  if (type === "Point") {
+    return geometry.coordinates
+  } else if (type === "Polygon") {
+    return geometry.coordinates[0][0]
+  } else {
+    throw new Error(`Badly formatted geometry. type=${type}`)
+  }
+}
+
+export const getCityFromProperties = (feature: IPointFeature | null) => feature?.properties?.city ?? feature?.properties?.municipality ?? feature?.properties?.locality ?? null
+export const getStreetFromProperties = (feature: IPointFeature | null) => (feature?.properties?.street ? feature?.properties?.name : null)
